Extract extractors to new module and add some docs
Pun fully intended.
