+++ autoendpoint/src/server/extractors/mod.rs
@@ -1,0 +1,6 @@
+//! Actix extractors (`FromRequest`). These extractors transform and validate
+//! the incoming request data.
+
+pub mod notification;
+pub mod subscription;
+pub mod token_info;
